+++ CHANGELOG.md
@@ -7,7 +7,6 @@
 * Fix #5186: Support for Pod uploads with big numbers
 * Fix #5221: Empty kube config file causes NPE
 * Fix #5281: Ensure the KubernetesCrudDispatcher's backing map is accessed w/lock
-* Fix #5293: Ensured the mock server uses only generic or JsonNode parsing
 * Fix #5298: Prevent requests needing authentication from causing a 403 response
 * Fix #5327: Ensured that the informer reconnect task terminates after client close
 * Fix #5113: Clashing package names in trigger model dependencies
@@ -17,6 +16,7 @@
 * Fix #5233: Generalized SchemaSwap to allow for cycle expansion
 * Fix #5262: all built-in collections will omit empty in their serialized form.
 * Fix #5287: Add an option to filter the files processed by the java-generator, based on a suffix allowlist
+* Fix #5293: Mock server supports KubernetesClientBuilder customization
 * Fix #5315: Introduced `kubernetes-junit-jupiter-autodetect` to use with [automatic extension registration](https://junit.org/junit5/docs/current/user-guide/#extensions-registration-automatic)
 * Fix #5339: `@PrinterColumn` annotation has configuration field for priority